Fix build warnings
- Remove unnecessary non-null assertion in BufferFactory.kt
- Suppress unchecked cast warnings in MutableDataBuffer.kt and AppleStreamingCompression.kt
- Fix redundant conversion in TransformedReadBufferTest.kt
- Use sourceSetTree to properly include commonTest in androidInstrumentedTest

+++ buffer/src/commonTest/kotlin/com/ditchoom/buffer/TransformedReadBufferTest.kt
@@ -26,7 +26,7 @@ class TransformedReadBufferTest {
         buffer.resetForRead()
 
         val transformed = TransformedReadBuffer(buffer, xorTransformer)
-        assertEquals((3.toInt() xor 0xFF).toByte(), transformed[2])
+        assertEquals((3 xor 0xFF).toByte(), transformed[2])
         assertEquals(0, transformed.position()) // Position unchanged
     }
